Likely final update.
This will probably be the last update for this repo, to be moved over to
Unity. This was started mosty to familiarize myself with some different
applications of C#, I think I have gone as far as I am willing here. I
am enjoying the actual progress of this project, and Monogame is the
limiting part of this project.